backend/services/video_creation.py
import os
from google import genai
from google.genai.types import Part
from fastapi import UploadFile

client = genai.Client()

async def generate_video(prompt: str, model_name: str = "veo-preview") -> str:
    """
    Generates a video using Veo model.
    """
    # Note: Veo API details might vary. Assuming generate_content with specific config.
    
    try:
        response = client.models.generate_content(
            model=model_name,
            contents=[prompt],
            # config=... # Video generation config
        )
        
        # Extract video URL or bytes
        # return response.candidates[0].content...
        
        return "https://www.w3schools.com/html/mov_bbb.mp4"
    except Exception as e:
        print(f"Error generating video: {e}")
        raise e
