requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
google-cloud-storage
jinja2
python-multipart
sqlalchemy
